Se cambio el servicio para que el nuevo valor del dromopunto sea de tipo decimal

+++ Backend/EventodromoRest/Mappers/DromopuntosMapper.cs
@@ -123,7 +123,7 @@ namespace EventodromoRest.Mappers
             return resumen;
         }
 
-        public int ActualizarValorDromoPuntos(int nuevoValor)
+        public int ActualizarValorDromoPuntos(decimal nuevoValor)
         {
             lock (DB)
             {
